requirements.txt
requests>=2.31
watchdog>=3.0
python-dotenv>=1.0
rich>=13.0
ollama>=0.2.0
mcp>=0.5.0
